Added annotations for DAO serialization/deserialization

+++ src/main/java/jcms/repository/PostRepository.java
@@ -3,8 +3,6 @@ package jcms.repository;
 import jcms.model.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface PostRepository extends JpaRepository<Post, BigInteger> {
-    Post findById(BigInteger id);
+public interface PostRepository extends JpaRepository<Post, Integer> {
+    Post findById(Integer id);
 }
