+++ src/main/java/jcms/repository/UserRepository.java
@@ -4,10 +4,8 @@ import jcms.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
-import java.math.BigInteger;
-
 @RepositoryRestResource(exported = false)
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     boolean existsByEmail(String email);
     boolean existsByUsername(String username);
     User findByUsername(String username);
